Remove blog page component
