+++ src/components/main/main-page.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 
 import MainSection from './sections/header/main-section'
 import ProjectSection from './sections/project/project-section'
-import BlogSection from './sections/blog/blog-section'
+import BlogSection from '../../containers/blog/blogPageLoader'
 import InfoSection from './sections/info/info-section'
 
 class MainPage extends Component {
